Move the registry into an extension with helpers for access.
First step in putting a lock around the registry mutation.
